Pull updates from wsl branch

+++ install.sh
@@ -29,29 +29,34 @@ for file in $(ls -a $path_to_self)
 do
     src="$path_to_self/$file"
     dest="$HOME/$file"
-    if [ -f $src ]
+    if [[ $file == "install.sh" ]]
     then
-        # $src is a non-directory file
-
-        # Check for existing files and symlinks
-        if [ -h "$dest" ]
+        echo -e "$info Skipping install script."
+    else
+        if [ -f $src ]
         then
-            # Remove existing symlink
-            echo -e  "$warn Found existing [$file] symlink. Removing."
-            rm "$dest"
+            # $src is a non-directory file
 
-        elif [ -f "$dest" ]
-        then
-            # backup existing
-            echo -e  "$info Found existing [$file]. Creating backup."
-            mv "$dest" "$BACKUP_LOCATION"
-        
-        fi
+            # Check for existing files and symlinks
+            if [ -h "$dest" ]
+            then
+                # Remove existing symlink
+                echo -e  "$warn Found existing [$file] symlink. Removing."
+                rm "$dest"
 
-        # Make the new link
-        echo -e  "$info Linking [$src] to [$dest]"
-        ln -s $src $dest 
-    fi 
+            elif [ -f "$dest" ]
+            then
+                # backup existing
+                echo -e  "$info Found existing [$file]. Creating backup."
+                mv "$dest" "$BACKUP_LOCATION"
+            
+            fi
+
+            # Make the new link
+            echo -e  "$info Linking [$src] to [$dest]"
+            ln -s $src $dest 
+        fi 
+    fi
 done
 
 
@@ -128,33 +133,80 @@ done
 
 
 # Fix up .tmux.conf POWERLINELOC
-powerlineloc=$(pip show powerline-status | grep Location | cut -d" " -f2)
+powerlineloc="$(pip show powerline-status | grep Location | cut -d" " -f2)"
 if [ ! -z "$powerlineloc" ]
 then
-    echo "$info Found powerline-status installed. Ensuring powerline is enabled"
-    sed -i "s/\#run-shell \"powerline/run-shell \"powerline/g" $HOME/.tmux.conf
-    sed -i "s@\#source /usr/local/lib/python@source /usr/local/lib/python@g" $HOME/.tmux.conf
+    
+    # Update powerline loc with the full path to the tmux conf
+    powerlineloc="$powerlineloc/powerline/bindings/tmux/powerline.conf"
+
+    # Assuming the link has been changed elsewhere, sed requires the path
+    # to the link target itself (it will overwrite the link without this)
+    # so it can edit the target.
+    if [ -h "$HOME/.tmux.conf" ]
+    then
+        TMUX_CONF="$(readlink -f $HOME/.tmux.conf)"
+    else
+        if [ -f "$HOME/.tmux.conf" ]
+        then
+            TMUX_CONF="$HOME/.tmux.conf"
+        fi
+    fi
+
+    echo -e "$info Found powerline-status installed. Ensuring powerline is enabled"
+    sed -i "s@\#run-shell \"powerline@run-shell \"powerline@g" $TMUX_CONF
+    sed -i "s@\#source POWERLINELOC@source "$powerlineloc"@g" $TMUX_CONF
 else
-    echo "$warn Failed to find powerline-status, Disabling tmux powerline"
-    sed -i "s/run-shell \"powerline/\#run-shell \"powerline/g" $HOME/.tmux.conf
-    sed -i "s@source /usr/local/lib/python@\#source /usr/local/lib/python@g" $HOME/.tmux.conf
+    echo -e "$warn Failed to find powerline-status, Disabling tmux powerline"
+    sed -i "s/run-shell \"powerline/\#run-shell \"powerline/g" $TMUX_CONF
+    sed -i "s@source /usr/local/lib/python@\#source /usr/local/lib/python@g" $TMUX_CONF
 fi
 
 
 # Fix up vimrc to enable vim pathogen if it is installed
 if [ -d "$HOME/.vim/autoload" ]
 then
+
+    # Grab real path to vimrc for sed
+    if [ -h "$HOME/.vimrc" ]
+    then
+        VIMRC="$(readlink -f $HOME/.vimrc)"
+    else
+        if [ -f "$HOME/.vimrc" ]
+        then
+            VIMRC="$HOME/.vimrc"
+        fi
+    fi
+
     # Found vim autoload dir, check for pathogen.vim
     if [ -f "$HOME/.vim/autoload/pathogen.vim" ]
     then
         echo -e "$info Found vim pathogen, ensuring it is enabled in vimrc"
-        sed -i "s/\"execute path/execute path/g" $HOME/.vimrc
+        sed -i "s/\"execute path/execute path/g" $VIMRC
     else
         echo -e "$warn Failed to find vim pathogen. Ensuring call is commented out"
-        sed -i "s/^execute path/\"execute path/g" $HOME/.vimrc
+        sed -i "s/^execute path/\"execute path/g" $VIMRC
     fi
 else
     # Pathogen not installed, ensure command is commented
     echo -e "$warn Failed to find vim autoload dir, pathogen not installed"
-    sed -i "s/^execute path/\"execute path/g" $HOME/.vimrc
+    sed -i "s/^execute path/\"execute path/g" $VIMRC
 fi
+
+
+# Write a simple gitconfig based on env-vars
+GCONF="$HOME/.gitconfig"
+if [ -e "$HOME/.gitconfig" ] # Handle existing config
+then
+    # Make a backup of the existing config
+    echo -e "$info Found existing .gitconfig, backing it up."
+    mv "$HOME/.gitconfig" "$BACKUP_LOCATION/.gitconfig_backup"
+fi
+# Now can write the new file
+echo -e "$info Writing new .gitconfig based on environment."
+echo "[User]" >> "$GCONF"
+echo "    name = $USER" >> "$GCONF"
+echo "    email = ""$USER""@""$HOST" >> "$GCONF"
+echo "[Core]" >> "$GCONF"
+echo "    editor = vim" >> "$GCONF"
+
